Refactor - moved components used in multiple pages in Create Facility to components folder
