+++ src/app/modules/create-facility/components/create-facility-container/create-facility-container.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Container, PageStateService } from 'moh-common-lib';
-import { createFacilityPageRoutes } from '../create-facility-page-routing';
-import { CREATE_FACILITY_PAGES } from '../create-facility-route-constants';
+import { createFacilityPageRoutes } from '../../create-facility-page-routing';
+import { CREATE_FACILITY_PAGES } from '../../create-facility-route-constants';
 import { HeaderService } from 'src/app/services/header.service';
