+++ src/app/modules/create-facility/components/review-applicant/review-applicant.component.spec.ts
@@ -1,23 +1,23 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ReviewFacilityComponent } from './review-facility.component';
-import { CoreBCPModule } from '../../../../core-bcp/core-bcp.module';
+import { ReviewApplicantComponent } from './review-applicant.component';
+import { CoreBCPModule } from '../../../core-bcp/core-bcp.module';
 import { RouterTestingModule } from '@angular/router/testing';
 
-describe('ReviewFacilityComponent', () => {
-  let component: ReviewFacilityComponent;
-  let fixture: ComponentFixture<ReviewFacilityComponent>;
+describe('ReviewApplicantComponent', () => {
+  let component: ReviewApplicantComponent;
+  let fixture: ComponentFixture<ReviewApplicantComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [ CoreBCPModule, RouterTestingModule ],
-      declarations: [ ReviewFacilityComponent ]
+      declarations: [ ReviewApplicantComponent ]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ReviewFacilityComponent);
+    fixture = TestBed.createComponent(ReviewApplicantComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
