+++ src/app/modules/create-facility/components/review-facility-bcp/review-facility-bcp.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ReviewFacilityBcpComponent } from './review-facility-bcp.component';
-import { CoreBCPModule } from '../../../../core-bcp/core-bcp.module';
+import { CoreBCPModule } from '../../../core-bcp/core-bcp.module';
 import { RouterTestingModule } from '@angular/router/testing';
 
 describe('ReviewFacilityBcpComponent', () => {
