+++ src/app/modules/create-facility/components/review-facility-bcp/review-facility-bcp.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { ReviewContainerComponent } from 'src/app/modules/core-bcp/components/review-container/review-container.component';
-import { CreateFacilityDataService } from '../../../services/create-facility-data.service';
-import { CREATE_FACILITY_PAGES } from '../../../create-facility-route-constants';
+import { CreateFacilityDataService } from '../../services/create-facility-data.service';
+import { CREATE_FACILITY_PAGES } from '../../create-facility-route-constants';
 
 @Component({
   selector: 'bcp-review-facility-bcp',
