+++ src/app/modules/create-facility/components/review-facility-mailing-info/review-facility-mailing-info.component.spec.ts
@@ -1,23 +1,23 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ReviewApplicantComponent } from './review-applicant.component';
-import { CoreBCPModule } from '../../../../core-bcp/core-bcp.module';
+import { ReviewFacilityMailingInfoComponent } from './review-facility-mailing-info.component';
+import { CoreBCPModule } from '../../../core-bcp/core-bcp.module';
 import { RouterTestingModule } from '@angular/router/testing';
 
-describe('ReviewApplicantComponent', () => {
-  let component: ReviewApplicantComponent;
-  let fixture: ComponentFixture<ReviewApplicantComponent>;
+describe('ReviewFacilityMailingInfoComponent', () => {
+  let component: ReviewFacilityMailingInfoComponent;
+  let fixture: ComponentFixture<ReviewFacilityMailingInfoComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [ CoreBCPModule, RouterTestingModule ],
-      declarations: [ ReviewApplicantComponent ]
+      declarations: [ ReviewFacilityMailingInfoComponent ]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ReviewApplicantComponent);
+    fixture = TestBed.createComponent(ReviewFacilityMailingInfoComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
