+++ src/app/modules/create-facility/components/review-facility-mailing-info/review-facility-mailing-info.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { ReviewContainerComponent } from 'src/app/modules/core-bcp/components/review-container/review-container.component';
-import { CreateFacilityDataService } from '../../../services/create-facility-data.service';
+import { CreateFacilityDataService } from '../../services/create-facility-data.service';
 import { getProvinceDescription } from 'moh-common-lib';
-import { CREATE_FACILITY_PAGES } from '../../../create-facility-route-constants';
+import { CREATE_FACILITY_PAGES } from '../../create-facility-route-constants';
 
 @Component({
   selector: 'bcp-review-facility-mailing-info',
