+++ src/app/modules/create-facility/components/review-facility/review-facility.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { ReviewContainerComponent } from 'src/app/modules/core-bcp/components/review-container/review-container.component';
-import { CreateFacilityDataService } from '../../../services/create-facility-data.service';
+import { CreateFacilityDataService } from '../../services/create-facility-data.service';
 import { getProvinceDescription } from 'moh-common-lib';
-import { setNotApplicable } from '../../../../core-bcp/models/helperFunc';
-import { CREATE_FACILITY_PAGES } from '../../../create-facility-route-constants';
+import { setNotApplicable } from '../../../core-bcp/models/helperFunc';
+import { CREATE_FACILITY_PAGES } from '../../create-facility-route-constants';
 
 @Component({
   selector: 'bcp-review-facility',
