+++ src/app/modules/create-facility/create-facility-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { CreateFacilityContainerComponent } from './create-facility-container/create-facility-container.component';
+import { CreateFacilityContainerComponent } from './components/create-facility-container/create-facility-container.component';
 import { createFacilityPageRoutes } from './create-facility-page-routing';
 import { CREATE_FACILITY_PAGES } from './create-facility-route-constants';
 import { SubmissionComponent } from './pages/submission/submission.component';
